requirements.txt
python-dotenv==1.0.1
requests==2.32.3
retell-sdk==4.0.0
fastapi==0.111.1